Revert heading de-increment
Based on discussion at https://github.com/manubot/rootstock/issues/273,
I have decided that the Pandoc convention of using h1 for sections
as well as the title is bizarre. It makes most sense for us to stick
with the Manubot convention of h2 for top-level sections like abstract.
This means numbered sections will always be problematic in HTML
output because they will be like 0.2.4. Therefore, revert --number-sections.